+++ src/log/engine_core.h
@@ -28,10 +28,10 @@ public:
     void BuildLocalCutMessage(LocalCutMsgProto* message);
     void OnNewFsmRecordsMessage(const FsmRecordsMsgProto& message);
 
-    bool LogTagToPrimaryNode(uint32_t tag, uint16_t* primary_node_id);
-    bool StoreLogAsPrimaryNode(uint32_t tag, std::span<const char> data, uint64_t* localid);
-    bool StoreLogAsBackupNode(uint32_t tag, std::span<const char> data, uint64_t localid);
-    void AddWaitForReplication(uint32_t tag, uint64_t localid);
+    bool LogTagToPrimaryNode(uint64_t tag, uint16_t* primary_node_id);
+    bool StoreLogAsPrimaryNode(uint64_t tag, std::span<const char> data, uint64_t* localid);
+    bool StoreLogAsBackupNode(uint64_t tag, std::span<const char> data, uint64_t localid);
+    void AddWaitForReplication(uint64_t tag, uint64_t localid);
 
     void DoStateCheck(std::ostringstream& stream) const;
